backend/requirements.txt
pyTelegramBotAPI
supabase
python-dotenv
postgrest